requirements.txt
Flask[async]
flask-login
Flask-Limiter
PyYAML
passlib
black
gunicorn